fix errore su angolo

+++ angolo.cpp
@@ -8,7 +8,7 @@ Angolo::Angolo(Coordinata& _c1,Coordinata& _c2,Coordinata& _c3, int _numeroframe
     numeroframe=_numeroframe;
 }
 
-Angolo::Angolo(double _azimut, double _zenit, double _numeroframe,int _maxmin) {
+Angolo::Angolo(double _azimut, double _zenit, int _numeroframe,int _maxmin) {
 	azimut = _azimut;
 	zenit = _zenit;
 	numeroframe = _numeroframe;
